Refactor showDelay condition to time delay after initialization.
