+++ src/rollup/display-conditions/watchers/validate-watchers.js
@@ -1,4 +1,7 @@
-/** @module pathfora/display-conditions/watchers/core/validate-watchers */
+/** @module pathfora/display-conditions/watchers/validate-watchers */
+
+// globals
+import { widgetTracker } from '../../globals/config';
 
 // display conditions
 import impressionsChecker from '../impressions/impressions-checker';
@@ -20,7 +23,14 @@ export default function validateWatchers (widget, cb) {
   }
 
   if (valid) {
-    showWidget(widget);
+    if (widget.displayConditions.showDelay) {
+      widgetTracker.delayedWidgets[widget.id] = setTimeout(function () {
+        showWidget(widget);
+      }, widget.displayConditions.showDelay * 1000);
+    } else {
+      showWidget(widget);
+    }
+
     widget.valid = false;
     cb();
