+++ src/rollup/widgets/init-widget.js
@@ -117,7 +117,13 @@ export default function initializeWidget (widget) {
 
     if (widget.watchers.length === 0 && !condition.showOnExitIntent) {
       if (widget.valid) {
-        showWidget(widget);
+        if (condition.showDelay) {
+          widgetTracker.delayedWidgets[widget.id] = setTimeout(function () {
+            showWidget(widget);
+          }, condition.showDelay * 1000);
+        } else {
+          showWidget(widget);
+        }
       }
     }
   });
